requirements.txt
Flask==3.1.0
numpy==2.2.3
Pillow==11.1.0
streamlit==1.43.2
tifffile==2025.2.18
